Checksum type: Move get_setting helper to more generic location
get_setting is now in archivematicaFunctions
